+++ clang-tools-extra/clang-tidy/tool/run-clang-tidy.py
@@ -82,7 +82,7 @@ def make_absolute(f, directory):
 def get_tidy_invocation(f, clang_tidy_binary, checks, tmpdir, build_path,
                         header_filter, allow_enabling_alpha_checkers,
                         extra_arg, extra_arg_before, quiet, config,
-                        line_filter):
+                        line_filter, fix):
   """Gets a command line for clang-tidy."""
   start = [clang_tidy_binary, '--use-color']
   if allow_enabling_alpha_checkers:
@@ -109,6 +109,8 @@ def get_tidy_invocation(f, clang_tidy_binary, checks, tmpdir, build_path,
       start.append('-quiet')
   if config:
       start.append('-config=' + config)
+  if fix:
+      start.append('--fix')
   start.append(f)
   return start
 
@@ -168,7 +170,7 @@ def run_tidy(args, tmpdir, build_path, queue, lock, failed_files):
                                      tmpdir, build_path, args.header_filter,
                                      args.allow_enabling_alpha_checkers,
                                      args.extra_arg, args.extra_arg_before,
-                                     args.quiet, args.config, args.line_filter)
+                                     args.quiet, args.config, args.line_filter, args.fix_tidy)
 
     proc = subprocess.Popen(invocation, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
     output, err = proc.communicate()
@@ -227,6 +229,7 @@ def main():
   parser.add_argument('files', nargs='*', default=['.*'],
                       help='files to be processed (regex on path)')
   parser.add_argument('-fix', action='store_true', help='apply fix-its')
+  parser.add_argument('-fix-tidy', action='store_true', help='apply fix-its')
   parser.add_argument('-format', action='store_true', help='Reformat code '
                       'after applying fixes')
   parser.add_argument('-style', default='file', help='The style of reformat '